fix(03-01): force int dtype in generatePandaEnumCols get_dummies (VAL-04)
pandas 2.x switched pd.get_dummies default dtype from uint8 to bool.
Downstream, the one-hot DataFrame is concat'd with int columns and the
combined .values array is passed to scipy.sparse.csr_matrix, which on
Py3.12+ rejects the resulting object dtype with
"scipy.sparse does not support dtype object".

Pin dtype=np.int8 in both the production helper (h2o/frame.py) and the
test util duplicate (tests/pyunit_utils/utilsPY.py) so the dtype stays
numeric and consistent across pandas versions. Fixes
pyunit_PUBDEV_5777_XGBoost_dense/sparse_comparisons on Py3.12/3.14.

+++ h2o-py/h2o/frame.py
@@ -5201,7 +5201,11 @@ def generatePandaEnumCols(pandaFtrain, cname, nrows, domainL):
             pass
     zeroFrame = pd.DataFrame(tempnp)
     zeroFrame.columns=cmissingNames
-    temp = pd.get_dummies(pandaFtrain[cname], prefix=cname, drop_first=False)
+    # pandas 2.x returns bool dtype from get_dummies by default. The result is concat'd
+    # with int columns and later passed to scipy.sparse.csr_matrix, which on Py3.12+
+    # rejects object dtype that mixed bool/int produce. Force int8 to keep the dtype
+    # numeric and consistent across pandas versions.
+    temp = pd.get_dummies(pandaFtrain[cname], prefix=cname, drop_first=False, dtype=np.int8)
     tempNames = list(temp)  # get column names
     colLength = len(tempNames)
     newNames = ['a']*colLength
